+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v0.7.1 (NOT YET RELEASED)
 * Additions
   * Added a function that prints information about the heap for debugging purposes (`debugHeap` in cx/utilities.go).
+  * Added capability of storing heap objects in a CX chain's program state.
 * Changes
   * Redesign of CX's garbage collector.
   * Changes to several functions that relied on allocating objects on the heap.